Default to "copy image" instead of "restore to pristine"
Decompressing the 7z image takes 9 seconds, while copying the image to
temperory directory takes less than one, on my laptop with SSD. Even
with spindle disks, it's likely that copying is way faster, considering
the effect of page cache.

Add a parameter to add_drive to "copy" the image, and use it in VirtTest
default devices. With that, default restore to disable.

This makes a cycle for each test and job much shorter.

Signed-off-by: Fam Zheng <famz@redhat.com>